Face_Recognition_app/requirements.txt
opencv-python
torch
torchvision
torchaudio
PIL
facenet_pytorch
pyqt5
sqlite3
Jetson.GPIO